requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-environ==0.12.0
djangorestframework==3.16.1
idna==3.11
requests==2.32.5
sqlparse==0.5.4
urllib3==2.5.0
